Dragging not working, but progress being made
