+++ wandb/board/ui/package.json
@@ -44,8 +44,10 @@
         "react-content-loader": "^3.0.1",
         "react-copy-to-clipboard": "^4.2.3",
         "react-dom": "^16.1.1",
+        "react-draggable": "^3.0.5",
         "react-easy-chart": "^1.0.0",
         "react-redux": "^5.0.3",
+        "react-resizable": "^1.7.5",
         "react-router": "^4.2.0",
         "react-router-dom": "^4.2.0",
         "react-router-redux": "^5.0.0-alpha",
